Use _static dir (from `html_static_path`)
This removes the dependency and works "just fine" for me at least.
I also think that bundling static html into the _static dir is better
semantically, but we could of course just add the old "backers_2014"
directory to the list (and remove "_static" because it's empty and
generates a warning).