fix docs title
As it appears to be hard to make Sphinx display a top-level
topic for backers without disrupting the docs' structure,
I've created a new entry in the TOC for the backers' list.
On one hand, it requires one more click to get to the list
of backers, but on the other hand, it should be more visible
in the first place.