Use html5lib as parser instead
That's way less painful as it parses stuff like a browser would and does
not nest everything if a td or p tag was not correctly closed.

Also do minor tweak adjustments for post-processing regarding html tags
and whitespaces. Trailing whitespaces on lines cause issues for the yaml
literal style.